Update ImageSegmentationBaseClass.m
added static methods to make lgraph a 1x5 cell array containing the layer graphs of each neural network. obj is no longer an input to the functions which define each of the networks, now it is just number of channels. added lines initializing an empty layer graph at the start of each nnet function.